MR-35. Fixed problems
