+++ src/Router.jsx
@@ -7,11 +7,11 @@ import { Route, Switch } from 'react-router-dom'
 import { routes } from './config/routes'
 
 // pages
-import { CategoryPage, Home, NewsDetailsPage, SearchPage } from './pages'
+import { CategoryPage, HomePage, NewsDetailsPage, SearchPage } from './pages'
 
 const Router = () => (
   <Switch>
-    <Route exact path={routes.home()} component={Home} />
+    <Route exact path={routes.home()} component={HomePage} />
     <Route path={routes.category()} component={CategoryPage} />
     <Route path={routes.newsDetails()} component={NewsDetailsPage} />
     <Route path={routes.search()} component={SearchPage} />
