+++ src/components/core/Header/Header.jsx
@@ -9,6 +9,9 @@ import Toolbar from '@material-ui/core/Toolbar'
 // components
 import { Search } from './libs/Search'
 
+// routes
+import { routes } from '../../../config/routes'
+
 // styles
 import { useStyles } from './styles'
 
@@ -19,7 +22,7 @@ const Header = () => {
     <AppBar className={classes.header}>
       <Toolbar>
         <Box flexGrow="1">
-          <NavLink to="/">
+          <NavLink to={routes.home()}>
             <img
               src="https://орфографика.рф/800/600/https/yt3.ggpht.com/a/AATXAJx5amDq3xleLn-gf7Kr02azA8uFO3_JVvO7dF3hRw=s900-c-k-c0xffffffff-no-rj-mo"
               alt="logo"
