+++ src/components/core/Sidebar/styles.js
@@ -24,4 +24,8 @@ export const useStyles = makeStyles(theme => ({
     flexGrow: 1,
     padding: theme.spacing(3),
   },
+  link: {
+    textDecoration: 'none',
+    color: 'inherit',
+  },
 }))
