+++ src/components/shared/index.js
@@ -1,0 +1,1 @@
+export * from './NewsList'
