+++ src/pages/Category/index.js
@@ -1,1 +1,1 @@
-export * from './CategoryPage'
+export * from './Category'
