+++ src/pages/Home/Home.jsx
@@ -1,0 +1,8 @@
+import React from 'react'
+
+// components
+import { Home } from '../../components/Home'
+
+const HomePage = () => <Home />
+
+export { HomePage }
