+++ src/pages/Home/index.js
@@ -1,1 +1,1 @@
-export * from './HomePage'
+export * from './Home'
