+++ src/pages/NewsDetails/NewsDetails.jsx
